+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.1.0 (Unreleased)
+# Changelog
+
+## v0.1.0 (Unreleased)
+
+- Initial release with basic functionality
+- Added `omni_machines` data source to list all machines in the Omni cluster
+- Added `omni_installation_media` data source to generate the schematic and pxe url
 
 FEATURES: